requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-froala-editor==4.3.0
django-honeypot==1.2.1
django-hosts==6.0
django-tenants==3.7.0
django-tenants-celery-beat==0.2.1
django-timezone-field==7.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools-scm==8.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tenant-schemas-celery==3.0.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
